Added timer and polymorphic_graph classes for use with ASEP
Small updates necessary to compile and run ASEP.  Created a polymorphic
graph class in order to have a universe for the ASEP particles to live
on.  On the other hand, probably not the most efficient way to store a
graph, so I gave it a specific name so in the future graph.h was
available.  The Timer class was a nice addition, by default it is
standard exponential jump times, but this is customizable via templates.